refactor(context): move all memory tests to benchmarks
Move shared references and per-message overhead tests from
memory.test.ts to context.bench.ts — these are performance
characterization checks, not unit tests.
