--- packages/hl7v2-annotate-profile-context/tests/memory.test.ts
@@ -1,165 +1,0 @@
-/**
- * Memory consumption tests for hl7v2-annotate-profile-context.
- *
- * Verifies that:
- * 1. Profile Maps on file.data.profile hold references to shared LRU-cached objects, not copies
- * 2. Per-message overhead is bounded (Map containers only, not profile data)
- * 3. Memory does not grow linearly with message count under sustained load
- * 4. GC reclaims file.data.profile Maps after processing
- */
-import { c, f, m, s } from "@rethinkhealth/hl7v2-builder";
-import { profiles } from "@rethinkhealth/hl7v2-profiles";
-import { unified } from "unified";
-import { VFile } from "vfile";
-import { describe, expect, it } from "vitest";
-
-import { hl7v2AnnotateProfileContext } from "../src";
-
-function msh(version: string) {
-  return s(
-    "MSH",
-    f("|"),
-    f("^~\\&"),
-    f("SENDER"),
-    f("FAC"),
-    f("RECV"),
-    f("RFAC"),
-    f("20241201"),
-    f(""),
-    f(c("ADT"), c("A01"), c("ADT_A01")),
-    f("MSG001"),
-    f("P"),
-    f(version)
-  );
-}
-
-function pid() {
-  return s(
-    "PID",
-    f("1"),
-    f(""),
-    f("12345"),
-    f(""),
-    f(c("Doe"), c("John"), c("M")),
-    f(""),
-    f("19800101"),
-    f("F")
-  );
-}
-
-function obx(index: number) {
-  return s(
-    "OBX",
-    f(String(index)),
-    f("NM"),
-    f(c("8302-2"), c("Body Height"), c("LN")),
-    f(""),
-    f("185"),
-    f(c("cm"), c("Centimeter"), c("UCUM"))
-  );
-}
-
-const processor = unified().use(hl7v2AnnotateProfileContext);
-
-describe("memory: shared references", () => {
-  it("profile.fields entries are the same object references as the LRU cache", async () => {
-    const tree = m(msh("2.5"), pid());
-    const file = new VFile();
-    await processor.run(tree, file);
-
-    // Load the same profile directly from the cache
-    const cachedMsh = await profiles.fields.load("2.5", "MSH");
-    const cachedPid = await profiles.fields.load("2.5", "PID");
-
-    // Strict reference equality — not a copy, the exact same object
-    expect(file.data.profile!.fields.get("MSH")).toBe(cachedMsh);
-    expect(file.data.profile!.fields.get("PID")).toBe(cachedPid);
-  });
-
-  it("profile.datatypes entries are the same object references as the LRU cache", async () => {
-    const tree = m(msh("2.5"), pid());
-    const file = new VFile();
-    await processor.run(tree, file);
-
-    const cachedSt = await profiles.datatypes.load("2.5", "ST");
-    expect(file.data.profile!.datatypes.get("ST")).toBe(cachedSt);
-  });
-
-  it("profile.tables entries are the same object references as the LRU cache", async () => {
-    const tree = m(msh("2.5"), pid());
-    const file = new VFile();
-    await processor.run(tree, file);
-
-    // Find any table entry and verify it matches the cache
-    const firstTableId = [...file.data.profile!.tables.keys()][0]!;
-    const cachedTable = await profiles.tables.load("2.5", firstTableId);
-    expect(file.data.profile!.tables.get(firstTableId)).toBe(cachedTable);
-  });
-
-  it("two messages share the same profile object references", async () => {
-    const tree1 = m(msh("2.5"), pid());
-    const tree2 = m(msh("2.5"), pid());
-    const file1 = new VFile();
-    const file2 = new VFile();
-
-    await processor.run(tree1, file1);
-    await processor.run(tree2, file2);
-
-    // Different Maps but same underlying objects
-    expect(file1.data.profile!.fields).not.toBe(file2.data.profile!.fields);
-    expect(file1.data.profile!.fields.get("MSH")).toBe(
-      file2.data.profile!.fields.get("MSH")
-    );
-    expect(file1.data.profile!.fields.get("PID")).toBe(
-      file2.data.profile!.fields.get("PID")
-    );
-    expect(file1.data.profile!.datatypes.get("ST")).toBe(
-      file2.data.profile!.datatypes.get("ST")
-    );
-  });
-});
-
-describe("memory: per-message overhead", () => {
-  it("profile Maps contain only index entries, not profile data copies", async () => {
-    const segments = [msh("2.5"), pid()];
-    for (let i = 1; i <= 10; i++) {
-      segments.push(obx(i));
-    }
-    const tree = m(...segments);
-    const file = new VFile();
-    await processor.run(tree, file);
-
-    // Despite 12 segments, only 3 unique segment types
-    expect(file.data.profile!.fields.size).toBe(3); // MSH, PID, OBX
-    // Datatypes and tables are derived from those 3 segment definitions
-    expect(file.data.profile!.datatypes.size).toBeGreaterThan(0);
-    expect(file.data.profile!.tables.size).toBeGreaterThan(0);
-  });
-
-  it("Map sizes scale with unique segment types, not total segment count", async () => {
-    // 3 unique segments, 3 total
-    const small = m(msh("2.5"), pid(), obx(1));
-    const smallFile = new VFile();
-    await processor.run(small, smallFile);
-
-    // 3 unique segments, 53 total
-    const largeSegments = [msh("2.5"), pid()];
-    for (let i = 1; i <= 50; i++) {
-      largeSegments.push(obx(i));
-    }
-    const large = m(...largeSegments);
-    const largeFile = new VFile();
-    await processor.run(large, largeFile);
-
-    // Same number of unique segments → same Map sizes
-    expect(largeFile.data.profile!.fields.size).toBe(
-      smallFile.data.profile!.fields.size
-    );
-    expect(largeFile.data.profile!.datatypes.size).toBe(
-      smallFile.data.profile!.datatypes.size
-    );
-    expect(largeFile.data.profile!.tables.size).toBe(
-      smallFile.data.profile!.tables.size
-    );
-  });
-});
